Define `oci.File`.
This type is currently just an alias for `v1.Image`, which is the unit of storage we use for files uploaded with `UploadFile[s]`.

This changes `static.NewFile` to create an `oci.File` (full image vs. just the layer previously), and `UploadFile` to use this (it's where the rest of the image came from).
