+++ pkg/oci/static/signature_test.go
@@ -27,91 +27,10 @@ import (
 	"github.com/sigstore/cosign/pkg/oci"
 )
 
-func TestNewFile(t *testing.T) {
-	payload := "this is the content!"
-	l, err := NewFile([]byte(payload), WithMediaType("foo"))
-	if err != nil {
-		t.Fatalf("NewFile() = %v", err)
-	}
-
-	t.Run("check size", func(t *testing.T) {
-		wantSize := int64(len(payload))
-		gotSize, err := l.Size()
-		if err != nil {
-			t.Fatalf("Size() = %v", err)
-		}
-		if gotSize != wantSize {
-			t.Errorf("Size() = %d, wanted %d", gotSize, wantSize)
-		}
-	})
-
-	t.Run("check media type", func(t *testing.T) {
-		wantMT := types.MediaType("foo")
-		gotMT, err := l.MediaType()
-		if err != nil {
-			t.Fatalf("MediaType() = %v", err)
-		}
-		if gotMT != wantMT {
-			t.Errorf("MediaType() = %s, wanted %s", gotMT, wantMT)
-		}
-	})
-
-	t.Run("check hashes", func(t *testing.T) {
-		wantHash, _, err := v1.SHA256(strings.NewReader(payload))
-		if err != nil {
-			t.Fatalf("SHA256() = %v", err)
-		}
-
-		gotDigest, err := l.Digest()
-		if err != nil {
-			t.Fatalf("Digest() = %v", err)
-		}
-		if !cmp.Equal(gotDigest, wantHash) {
-			t.Errorf("Digest = %s", cmp.Diff(gotDigest, wantHash))
-		}
-
-		gotDiffID, err := l.DiffID()
-		if err != nil {
-			t.Fatalf("DiffID() = %v", err)
-		}
-		if !cmp.Equal(gotDiffID, wantHash) {
-			t.Errorf("DiffID = %s", cmp.Diff(gotDiffID, wantHash))
-		}
-	})
-
-	t.Run("check content", func(t *testing.T) {
-		comp, err := l.Compressed()
-		if err != nil {
-			t.Fatalf("Compressed() = %v", err)
-		}
-		defer comp.Close()
-		compContent, err := io.ReadAll(comp)
-		if err != nil {
-			t.Fatalf("ReadAll() = %v", err)
-		}
-		if got, want := string(compContent), payload; got != want {
-			t.Errorf("Compressed() = %s, wanted %s", got, want)
-		}
-
-		uncomp, err := l.Uncompressed()
-		if err != nil {
-			t.Fatalf("Uncompressed() = %v", err)
-		}
-		defer uncomp.Close()
-		uncompContent, err := io.ReadAll(uncomp)
-		if err != nil {
-			t.Fatalf("ReadAll() = %v", err)
-		}
-		if got, want := string(uncompContent), payload; got != want {
-			t.Errorf("Uncompressed() = %s, wanted %s", got, want)
-		}
-	})
-}
-
 func TestNewSignatureBasic(t *testing.T) {
 	payload := "this is the content!"
 	b64sig := "b64 content=="
-	l, err := NewSignature([]byte(payload), b64sig, WithMediaType("foo"))
+	l, err := NewSignature([]byte(payload), b64sig, WithLayerMediaType("foo"))
 	if err != nil {
 		t.Fatalf("NewSignature() = %v", err)
 	}
@@ -265,7 +184,7 @@ func TestNewSignatureBasic(t *testing.T) {
 
 func TestNewAttestationBasic(t *testing.T) {
 	payload := "this is the content!"
-	l, err := NewAttestation([]byte(payload), WithMediaType("foo"))
+	l, err := NewAttestation([]byte(payload), WithLayerMediaType("foo"))
 	if err != nil {
 		t.Fatalf("NewSignature() = %v", err)
 	}
